requirements.txt
PyYAML
pandas
flask
dnspython
fastapi
pymongo
python-dotenv
uvicorn
websockets
xgboost
streamlit
scipy
numpy
seaborn
scikit-learn
geopandas
statsmodels
optuna
dvc
dill
mlflow
-e .